+++ ProjectEuler/ProjectEuler/Problems/Problem1.h
@@ -8,11 +8,11 @@ namespace Problems{
 	class Problem1 : public Problem{
 	private:
 		int limit;
-		int solution1();
-		int solution2();
+		__int64 solution1();
+		__int64 solution2();
 	public:
 		Problem1(void);
-		int run(void);
+		__int64 run(void);
 	};
 
 }
